requirements.txt
anyio==3.6.1
APScheduler==3.9.1
asgiref==3.5.2
cachetools==5.2.0
certifi==2022.6.15
charset-normalizer==2.1.0
Django==4.0.6
django-cors-headers==3.13.0
django-extensions==3.2.0
djangorestframework==3.13.1
fastapi==0.85.0
flake8==5.0.4
google-api-core==2.8.2
google-api-python-client==2.52.0
google-auth==2.9.0
google-auth-httplib2==0.1.0
googleapis-common-protos==1.56.3
httplib2==0.20.4
idna==3.3
mccabe==0.7.0
oauth2client==4.1.3
protobuf==4.21.2
psycopg2==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.9.1
pydantic==1.10.2
pyflakes==2.5.0
pyparsing==3.0.9
python-dotenv==0.21.0
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
requests==2.28.1
rsa==4.8
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.2
starlette==0.20.4
typing_extensions==4.3.0
tzdata==2022.1
tzlocal==4.2
uritemplate==4.1.1
urllib3==1.26.10
